+++ sdk/expo/CHANGELOG.md
@@ -5,7 +5,7 @@
 - [minor] License Update
 - [NA] Fixing the failing cases due to types/node version update
 - Android
-  - [minor] Bundling `MoEFireBaseMessagingService` in Plugin's manifest to reduce integration steps for developers using FCM Push.
+  - [minor] Compatibility changes for "Bundling `MoEFireBaseMessagingService` in Plugin's manifest to reduce integration steps for developers using FCM Push." in core plugin
 
 # 25-08-2025